# internal/handler/system/routes.go
package system

import (
	"encoding/json"
	"errors"
	"github.com/go-chi/chi/v5"
	handlerutils "github.com/shuTwT/nex-api/internal/pkg/utils"
	servicesystem "github.com/shuTwT/nex-api/internal/service/system"
	"github.com/shuTwT/nex-api/pkg/domain/model"
	"net/http"
)

type Handler struct {
	service *servicesystem.Service
	mux     chi.Router
}

type responseEnvelope = model.EnvelopeResp
type initializeData = model.SystemInitializeResp

func NewHandler(service *servicesystem.Service) http.Handler {
	handler := &Handler{service: service, mux: chi.NewRouter()}
	handler.registerRoutes(handler.mux)
	return handler
}

func RegisterRoutes(r chi.Router, service *servicesystem.Service) error {
	if r == nil {
		return errors.New("system: route mux is nil")
	}
	if service == nil {
		return errors.New("system: service is nil")
	}
	(&Handler{service: service}).registerRoutes(r)
	return nil
}

func (h *Handler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	if h.service == nil {
		writeJSON(w, http.StatusInternalServerError, responseEnvelope{Success: false, Error: "system_unavailable"})
		return
	}
	h.mux.ServeHTTP(w, r)
}

func (h *Handler) registerRoutes(r chi.Router) {
	r.Get("/api/system/initialized", h.initialized)
	r.Post("/api/system/initialize", h.initialize)
}

// initialized handles GET /api/system/initialized.
//
// @Summary GET /api/system/initialized
// @ID system_initialized_route_get
// @Tags system
// @Produce json
// @Success 200 {object} main.SwaggerEnvelope{data=bool}
// @Failure 400 {object} main.SwaggerEnvelope
// @Failure 401 {object} main.SwaggerEnvelope
// @Failure 500 {object} main.SwaggerEnvelope
// @Router /api/system/initialized [get]
func (h *Handler) initialized(w http.ResponseWriter, r *http.Request) {
	initialized, err := h.service.Initialized(r.Context())
	if err != nil {
		writeJSON(w, http.StatusInternalServerError, responseEnvelope{Success: false, Error: "initialization_status_failed"})
		return
	}
	writeJSON(w, http.StatusOK, responseEnvelope{
		Success: true,
		Data:    map[string]bool{"initialized": initialized},
	})
}

// initialize handles POST /api/system/initialize.
//
// @Summary POST /api/system/initialize
// @ID system_initialize_route_post
// @Tags system
// @Accept json
// @Produce json
// @Param body body model.SystemInitializeReq true "request payload"
// @Success 200 {object} main.SwaggerEnvelope{data=model.SystemInitializeResp}
// @Failure 400 {object} main.SwaggerEnvelope
// @Failure 401 {object} main.SwaggerEnvelope
// @Failure 500 {object} main.SwaggerEnvelope
// @Router /api/system/initialize [post]
func (h *Handler) initialize(w http.ResponseWriter, r *http.Request) {
	var request model.SystemInitializeReq
	if err := handlerutils.DecodeJSON(r, &request); err != nil {
		writeJSON(w, http.StatusBadRequest, responseEnvelope{Success: false, Error: "invalid_request"})
		return
	}
	admin, err := h.service.Initialize(r.Context(), request)
	if errors.Is(err, servicesystem.ErrAlreadyInitialized) {
		writeJSON(w, http.StatusBadRequest, responseEnvelope{Success: false, Error: "already_initialized"})
		return
	}
	if errors.Is(err, servicesystem.ErrInvalidInitialize) {
		writeJSON(w, http.StatusBadRequest, responseEnvelope{Success: false, Error: "invalid_request"})
		return
	}
	if err != nil {
		writeJSON(w, http.StatusInternalServerError, responseEnvelope{Success: false, Error: "initialization_failed"})
		return
	}
	writeJSON(w, http.StatusCreated, responseEnvelope{
		Success: true,
		Data: initializeData{
			ID: admin.ID, Email: admin.Email, Username: admin.Username, Role: admin.Role,
			Credits: admin.Credits,
		},
	})
}

func writeJSON(w http.ResponseWriter, status int, value interface{}) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	return json.NewEncoder(w).Encode(value)
}