# openapi/swagger/swagger.yaml
basePath: /
definitions:
...
  main.SwaggerEnvelope:
    properties:
      data: {}
      error:
        type: string
      pagination:
        $ref: '#/definitions/main.SwaggerPagination'
      success:
        type: boolean
    type: object
  main.SwaggerPagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  model.SystemInitializeReq:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  model.SystemInitializeResp:
    properties:
      credits:
        type: integer
      email:
        type: string
      id:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
  upload.Metadata:
    properties:
      filename:
        type: string
      size:
        type: integer
      type:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: Contract for the Nex HTTP API.
  license:
    name: MIT
    url: https://opensource.org/license/mit
  title: Nex API
  version: 1.0.0
paths:
...
  /api/system/initialize:
    post:
      consumes:
      - application/json
      operationId: system_initialize_route_post
      parameters:
      - description: request payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.SystemInitializeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/model.SystemInitializeResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: POST /api/system/initialize
      tags:
      - system
  /api/system/initialized:
    get:
      operationId: system_initialized_route_get
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  type: boolean
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/system/initialized
      tags:
      - system
...
  /api/upload:
    post:
      consumes:
      - multipart/form-data
      operationId: upload_route_post
      parameters:
      - description: uploaded file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.SwaggerEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/upload.Metadata'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      security:
      - SessionCookie: []
      summary: POST /api/upload
      tags:
      - upload
  /api/upload/{filename}:
    get:
      operationId: upload_filename_route_get
      parameters:
      - description: filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.SwaggerEnvelope'
      summary: GET /api/upload/{filename}
      tags:
      - upload
...
securityDefinitions:
  ApiTokenAuth:
    in: header
    name: Authorization
    type: apiKey
  CronSecretAuth:
    in: header
    name: Authorization
    type: apiKey
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"